source/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "simulation"
version = "0.1.0"
dependencies = [
    "pyyaml",
    "neuron",
    "LFPy",
    "numba",
    "SpikeInterface[full]",
    "MEArec",
    "numpy",
    "xarray",
    "pytest",
]
description = "Simulations for a neuron tracking study."
authors = [{name = "Oliver Shetler", email = "cos2112@cumc.columbia.edu"}]
readme = "README.md"

[tool.setuptools.packages.find]
exclude = ["config", "input", "output"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"').",
    "serial",
    "smoke: marks tests as minimal smoke tests meant for quick use during development."
]